Work around ambiguities in BSON::DBRef 'ref'
Certain versions of Moo change when/how deferred subroutines work.  The
latest breakage was noticed with Moo 2.002002.

This commit renames attribute 'ref' to '_ref' to avoid most collisions,
calls 'new' during compilation to force it to compile early, then aliases
'ref' back to '_ref' so it can still be used as a familiar accessor.